Create an UnmarshalUtil istance for each spawned thread
Should solve "Unmarshalling exception: FWK005 parse may not be called while
parsing." errors. The problem is that a singleton was used, but was not thread
safe (because of classes and methods used). Then when the parsing was performed
at the same time by two threads an exception would be thrown.